requirements.txt
sentence-transformers
faiss-cpu
rank_bm25
numpy
PyMuPDF
torch
mcp
